Logged in user updates a contacts and delete  contact

+++ controllers/contactController.js
@@ -46,6 +46,10 @@ const updateContact =asyncHandler(async(req, res) => {
         res.status(404)
         throw new Error('Contact not found')
     }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("User don't have permission to update")
+    }
     const updatedContent = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -63,7 +67,10 @@ const deleteContact = asyncHandler(async(req, res) => {
         res.status(404)
         throw new Error('Contact not found')
     }
-
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("User don't have permission to delete")
+    }
     await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json(contact)
 })
